fix: warnings under Windows+MSVC+x86 (#4515)
Remove any remaining warnings under Windows+MSVC+x86, now the examples,
tests and benchmarks compile cleanly in 32-bit mode. I think we are
actually using the correct types, not just silencing warnings with
`static_cast<>` too.

To enable `/WX` (treat warnings as errors) I had to disable warnings
around an Abseil include, there is a TODO for a better fix.

+++ google/cloud/spanner/benchmarks/benchmarks_config.cc
@@ -113,9 +113,9 @@ google::cloud::StatusOr<Config> ParseArgs(std::vector<std::string> args) {
          c.maximum_clients = std::stoi(v);
        }},
       {"--table-size=",
-       [](Config& c, std::string const& v) { c.table_size = std::stol(v); }},
+       [](Config& c, std::string const& v) { c.table_size = std::stoi(v); }},
       {"--query-size=",
-       [](Config& c, std::string const& v) { c.query_size = std::stol(v); }},
+       [](Config& c, std::string const& v) { c.query_size = std::stoi(v); }},
 
       {"--use-only-stubs",
        [](Config& c, std::string const&) { c.use_only_stubs = true; }},
